requirements.txt
pyheif
Pillow
ExifRead
dateparser